requirements.txt
alabaster
annotated-types==0.7.0
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgi-redis==1.1.0
asgiref==3.8.1
astroid==3.1.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autobahn
Automat==22.10.0
Babel==2.14.0
black==23.12.1
boto3==1.34.119
botocore==1.34.143
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
channels==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
coverage==7.4.0
crispy-bootstrap5==2023.10
cryptography==42.0.5
cssbeautifier==1.15.1
daphne==3.0.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
Django==4.2.9
django-allauth==0.60.1
django-anymail==10.3
django-apscheduler==0.6.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==4.3.1
django-coverage-plugin==3.1.0
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-ipware==7.0.1
django-jazzmin==3.0.0
django-model-utils==4.3.1
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-redis==5.4.0
django-storages==1.14.3
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.14.5
djlint==1.34.1
docutils==0.20.1
drf-spectacular==0.27.1
EditorConfig==0.12.4
exceptiongroup==1.2.1
executing==2.0.1
factory-boy==3.3.0
Faker==25.0.0
filelock==3.14.0
flake8==7.0.0
flake8-isort==6.1.1
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hiredis==2.3.2
hpack==4.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
hyperframe==6.0.1
hyperlink==21.0.0
identify==2.5.36
idna==3.7
imagesize==1.4.1
importlib_metadata==7.1.0
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jsbeautifier==1.15.1
json5==0.9.25
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
livereload==2.6.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
msgpack-python==0.5.6
mypy==1.7.1
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==24.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==8.13.35
pillow==10.2.0
platformdirs==4.2.1
pluggy==1.5.0
pre-commit
priority==1.3.0
prompt-toolkit==3.0.43
psycopg==3.1.17
psycopg-binary==3.1.18
psycopg-c==3.1.17
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyOpenSSL==24.1.0
pytest==7.4.4
pytest-django==4.7.0
pytest-sugar==0.9.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ipware==3.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rpds-py==0.18.0
s3transfer==0.10.2
service-identity==24.1.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
sqlparse==0.5.0
stack-data==0.6.3
termcolor==2.4.0
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.12.4
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
Twisted==24.3.0
txaio==23.1.1
types-pytz==2024.1.0.20240417
types-PyYAML==6.0.12.20240311
types-requests==2.31.0.20240406
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.3
virtualenv==20.26.1
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
whitenoise==6.6.0
zipp==3.18.1
zope.interface==6.4.post2
fcm-django